(fix & refactor) Faster and more scalable frame handling
Problem to fix: after many requests on a connection, the read-frame got slower,
and worse, system eventually died due to no memory available.

Fix: Open connections no longer keet reference to all the closed streams. It
means that they can now be GC'd. What is not in the list is classified as either
idle or closed.

The search from stream ID to stream object is factored out, and improved.

The logic for closing streams is now moved to the send frame/read frame to make
sure that the state change is enforced in every case (it was not).

Details:

- Add methods for :closed state for get-state, PEER-RESETS-STREAM,
APPLY-WINDOW-SIZE-INCREMENT

- Remove PEER-OPENS-HTTP-STREAM. The login is now in FIND-HTTP-STREAM-BY-ID
based on frame type object.

- Add IS-OUR-STREAM-ID so that the connection object can specify if it wants
even or odd stream IDs

- Factor stream closing logic to CLOSE-HTTP2-STREAM and cut the closed streams
from the streams list in it.

- Check initial peer window size setting as required in the spec.

- DEFINE-FRAME-TYPE has a new parameter, NEW-STREAM-STATE.

- CLOSE-HTTP2-STREAM is called automatically from READ-XXX-FRAME and
WRITE-XXX-FRAME when END-STREAM is set.

- Frame type now contains additional slots used by the FIND-HTTP-STREAM-BY-ID.