(fix & refactor) Faster and more scalable frame handling
Problem to fix: after many requests on a connection, the read-frame got slower,
and worse, system eventually died due to no memory available.

Fix: Open connections no longer keet reference to all the closed streams. It
means that they can now be GC'd. What is not in the list is classified as either
idle or closed.

The search from stream ID to stream object is factored out, and improved.

The logic for closing streams is now moved to the send frame/read frame to make
sure that the state change is enforced in every case (it was not).